[MIN-153] Remove getLiquidationPoolParams.
